test2/kafka-docker/python-script/requirements.txt
kafka-python==2.0.2
zeep>=4.1.0